+++ src/unit/code.rs
@@ -1,5 +1,5 @@
-pub const MIN_DISTANCE: f64 = 0.001;
+pub const MIN_DISTANCE: f32 = 0.001;
 pub const INTO_END: u32 = 1;
 pub const DIE:u32=2;
 
-pub const BULLET_HIT_DISTANCE:f64=0.001;
+pub const BULLET_HIT_DISTANCE:f32=0.001;
